mention bundle_gemfile in readme
Update the README to mention that `BUNDLE_GEMFILE` has to be set in
order for the tests to run correctly, which reflects how tests are
run in Travis.

When the tests are run in Travis, the environment variable
`BUNDLE_GEMFILE` is set to one of the gemfiles in the `gemfiles/`
directory.

Running the `bundle exec rake spec` command without a `BUNDLE_GEMFILE`
causes all the tests to fail with
`LoadError: cannot load such file -- rails`